define and implement in memory persist store, generate and add proto model

+++ src/HamustroNClient/Core/DateTimeEx.cs
@@ -5,11 +5,11 @@ namespace HamustroNClient.Core
 {
     public static class DateTimeEx
     {
-        public static int GetEpochUtc(this DateTime dt)
+        public static ulong GetEpochUtc(this DateTime dt)
         {
             var t = dt - new DateTime(1970, 1, 1);
             
-            return (int) t.TotalSeconds;
+            return (ulong) t.TotalSeconds;
         }
     }
 }
